Split templates.js into individual files
