+++ tools/publish/lib/index.js
@@ -1,6 +1,6 @@
 const { scanVault, buildLinkMap, scanAttachments, slugify, mapFolder } = require('./scanner');
 const { processContent, extractSections, resolveWikiLinks, filterSections, stripDataview, stripLeadingH1, renderRelationships, relativePath, escapeHtml, resolveImageEmbeds } = require('./processor');
-const templates = require('./templates');
+const templates = require('./templates/index');
 const { build } = require('./build');
 
 module.exports = {
