+++ tools/publish/lib/templates/index.js
@@ -1,0 +1,26 @@
+const { DIR_LABELS, cssPath, rootPath, baseShell, stubBadge, TYPE_BADGE_FIELDS, metadataBadgesFor, portraitImg } = require('./base');
+const { generateNav } = require('./nav');
+const { pcTemplate } = require('./pc');
+const { npcTemplate } = require('./npc');
+const { locationTemplate } = require('./location');
+const { wikiTemplate } = require('./wiki');
+const { indexTemplate } = require('./index-page');
+const { landingTemplate } = require('./landing');
+
+module.exports = {
+  DIR_LABELS,
+  cssPath,
+  rootPath,
+  baseShell,
+  stubBadge,
+  TYPE_BADGE_FIELDS,
+  metadataBadgesFor,
+  portraitImg,
+  generateNav,
+  pcTemplate,
+  npcTemplate,
+  locationTemplate,
+  wikiTemplate,
+  indexTemplate,
+  landingTemplate,
+};
