+++ tools/publish/lib/templates/wiki.js
@@ -1,0 +1,22 @@
+const { escapeHtml } = require('../processor');
+const { baseShell, cssPath, rootPath, stubBadge, metadataBadgesFor, portraitImg } = require('./base');
+
+function wikiTemplate(page, processedContent, navFor, config, imageMap) {
+  const fm = page.frontmatter;
+  const badges = metadataBadgesFor(fm);
+  const portrait = portraitImg(fm, page.outputPath, imageMap || {});
+
+  const content = `<h1 class="page-title">${escapeHtml(page.title)}${stubBadge(fm)}</h1>\n${portrait}\n${badges}\n${processedContent.html}\n${processedContent.relationships}`;
+
+  return baseShell({
+    title: page.title,
+    siteTitle: config.siteTitle,
+    cssHref: cssPath(page.outputPath),
+    navHtml: navFor(page.outputPath),
+    rootHref: rootPath(page.outputPath),
+    content,
+    footer: config.footer,
+  });
+}
+
+module.exports = { wikiTemplate };
